bringup: share bringup method targe and agument by two steps
First we formalize now a kdevops provisioner as for one which can
bring up nodes for us, can optionally setup ssh for us, and handles
destroying nodes.

Prior to this patch we defined provisioning as a two step process:

1) ssh update
2) devconfig playbook is run

However we implicitly were issuing each provisioning script. So
let us augment this first by defining the provisionin method first.
The devconfig playbook target is also common, so no need to
keep that separate. By sharing the devconfig playbook run as a shared
bringup goal we also implicitly fix bare metal bringup by ensuring we
run the devconfig playbook once at bringup.

We also enhance this with a new step:

KDEVOPS_BRING_UP_DEPS_EARLY

This variable can be augmented by new features for so to be able
to install things using the devconfig playbook before the general
devconfig playbook is run.

The diff stat reveals sharing is best. It will allow us later
to say, for example add ktest support or xfstests-bld, if we want to in a
much easier smoother way, as yet another provisioner.

This should pretty much be a no-op for most deployments, but enhances
bare metal setup so to ensure we run the devconfig playbook. It also
allows makes it easier to add new tools which may be optional which
we want installed early.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>