bringup: move journal-server setup early
Now that we have done the cleaning up of the provisioners and have
defined an early target task which we can piggy the back on top before
the devconfig playbook is run, move the journal server setup early.

This does not make any functional changes to the file to make it easier
to review the functional changes, those will be done next.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>