journal-server: fix by adjusting ordering
Now that we have done all the legwork to address ordering for bringup
and allow for items to be setup / installed with the devconfig playbook
for specific features, prior to running all generic devconfig tasks,
move the systemd remote journal server and client setup to run early.

Fix also ensuring we only try to setup the server once.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>