bringup: fix the soft iWARP bringup
It seemed like this worked for me at some point, but today it was
failing to start for me. In any case, removing the ACTION= filter seemed
to fix it for me.

Also, after reloading the udev rules, trigger them to be run so the
siw interface is immediately started.

Signed-off-by: Jeff Layton <jlayton@kernel.org>