journal-server: enhance ordering
The journal-server target is a localhost work item, so use the
provided variable for that. Then the journal-ln target needs to be
run at the very end. Make this explicit.

Note that running 'make journal-ln' is safe many times over, you just
don't want to run in in between a 'make fstests-baseline'as the
fstests watchdog may be reading a file. So we will later extend fstests
to ensure we run it, so to also delete stale journals.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>