journal-server: enforce non-SSL and clarify this with Kconfig
The systemd remote journal support was missing one last item: disabling
https. But we want to be explicit about this so add a new Kconfig entry
for it. Also provide some guidance for cloud users what they might want
to do if they want to enable this in the future.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>