fstests_watchdog.py: get kernel using remote journal
The whole point to enabling the systemd remote journal stuff
was the ability to not have to ssh to the remote note to query
for information. However the fstests_watchdog.py still does ssh
to the target node for two things:

1) Get current kernel
2) Get current time

This fixes 1) by using the journal and greping for the first line
with "Linux version". In order to do this we have to augment the
files we parse, so we also parse older files, if they exist.

This should speed up the fstests_watchdog.py considerably, we
now just need to address 2) and the watchdog should be *super fast*.

This also enhances the watchog to query all the journals for when
looking for a current fstests run. This was just missing. Note that
in the future it would be nice to have systemd journalctl just do
this sort of stuff for us, so for now we open code all this. What
do we want from systemd:

a) A parameter which just gets the last known kernel used
b) A parameter which gets the last time value from the journal
c) A shortcut for remote journals, so we don't have to do our
   own list of expected journals, with files like:

   - host.journal
   - host@*.journal

A kernel debugging setup should not need such high paranoid setup
to require root too. Root is required for containers with journalctl -M,
we want somethin similar and lightweight for remote machines for which
the user already has access to the remote directory for.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>