fstests_watchdog.py: use local host system time
As noted in the prior commit the last part was figuring out a way to
get current time. We'll use the host's own system time. We also have
to skip the host check for the process running "check" on /proc and so
we simply treat fstestsdone tag as an indication the tests have been
completed and everything is OK.

This completes the removal of having to use ssh for the fstests watchdog
and also allows us to capture now hung tests even when the host has
crashed completely.

We may later want to use and setup NTP, but this is a great improvement.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>