fstests_watchdog.py: provide a fallback for kernel version
If we cannot infer the kernel using the remote systemd journal files,
it is likely they already rotated and we can't get to the first journal
as 'make journal-ln' is stale.

There are two ways to fix this, the long term hard way is to avoid using
hostnames or force the hostnames on the journal. I coud not resolve that
last time I tried, and henceh the 'make journal-ln' as a post bringup
setup. The easier thing to do is to simply infer the kernel based on the
target kernel configuration and if the user didn't configure a kernel we
simply use "Distro-kernel".

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>