devconfig: add new systemd-timesyncd support
The new systemd-remote-journal support has allowed us to allow for
a huge simplification of the fstests / blktests watchdog. One of the
things that it did was that now that journals are local to the host
we can query the journals fast. There were a few lingering cases
where we were still ssh'ing in though:

  a) to get the kernel version
  b) get the current time

While a) and providing an infferred value in case the we cannot read it
works, for b) we get odd numbers for the status runtime of a test in
cases where we have huge drifts between the host and client as we now
rely on the host's notion of time for what is the current time. To fix
these odd drifts this adds systemd-timesyncd support so that we can set
the timezone and enable NTP on both the host and clients.

We then enhance systemd-remote-journal support by ensuring enabling it
also enables systemd-timesyncd. We ensure we set up timezone and NTP
prior to systemd-remote-journal by including the work for systemd-timesyncd
before the remote journal support.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>