bootlinux: add support for KERNELRELEASE
Add support for KERNELRELEASE make parameter to avoid UTS changes.
This is when LOCALVERSION is used, a '+' at the end of the kernel
release string is added, indicating the repository is dirty and has
local changes. Therefore, this can be used to specify the final string
value returned by the 'uname -r' or 'uname --kernel-version' command
(without the '+').

More information:
https://lore.kernel.org/lkml/CAK7LNAQmXwpEpY84-fdLiwALSFLUEkfsOHejbqy33c6ceM=4Tw@mail.gmail.com/T/#m273f4ae506aa71b3db91b269e5b2a18461d279e9

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>