guestfs: Specify host ISA to virt-builder
According to documentation, the default architecture for "virt-
builder" isn't always the same as the host's ISA.

Specifically for guestfs/libvirt, though, we want to ensure that the
host and guest ISA to match in order to avoid ISA emulation, which
is slow.

It shouldn't be difficult to customize this later on, if someone
wants to run libvirt guests under emulation.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>