guestfs: Enable destruction of guests with NVRAM
The default guest configuration on ARM systems includes a virtual
NVRAM device. However, "virsh undefine" won't touch such guests
without an explicit command-line option that enables destruction of
that device and its contents.

Note that virsh(1) says:

> --nvram and --keep-nvram specify accordingly to delete or keep
> nvram (/domain/os/nvram/) file.

In other words, "virsh undefine" is supposed to delete that file,
but currently does not. This is a known bug.

Link: https://lists.libvirt.org/archives/list/users@lists.libvirt.org/thread/MG3BFEOO7DUX6APTL7NXHOVZE4K2TO22/
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>