libvirt_user: Run Fedora-specific libvirt_user set up on Fedora systems
To run kdevops as a regular user, typically you set the
LIBVIRT_CONFIGURE Kconfig option. This is supposed to enable your
user account to run libvirt without needing root permission.

On my Fedora 39 system, however, this was not working. Virtual
networking would not start automatically, and when I tried to bring
it up manually, it threw errors.

I discovered that there is a libvirt_user set-up role for Fedora
that was not getting run. The set-up role for Red Hat excludes an
important step.

Add logic to invoke the Fedora action instead of the Red Hat action
on Fedora systems. As an additional clean up, replace "import_tasks"
with "include_tasks" because that seems to be more reliable for
Ansible install-deps actions.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>