nfsd: rework how storage is allocated to the LVM VG
Currently we require the user to put in two different path strings, and
they have to reset that if the configuration changes. This is limiting,
since we need to add the ability to suck in a 3rd disk.

Rework it to take a device prefix string and a count. "count" number of
devices will be added to the VG, starting at index 1.

Signed-off-by: Jeff Layton <jlayton@kernel.org>