gitr: Change the default mode of root directories under test
The default value of data_user is "vagrant". We should replace
that some day, but that looks a bit hairy to tackle.

When testing in a guestfs environment, the "vagrant" user ID is
not defined on target hosts, resulting in immediate test failures
(at least for the gitr workflow).

Address this by creating test file systems so they are owned by root,
and change the default mode of the root directory to 1777. This
enables any user to create files and directories on the test
file system.

Add a playbook parameter so that playbook consumers can set a
specific mode if they require it.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>