scripts/bringup_guestfs.sh: allow libvirt user to modify extra disks
When a system session is used libvirt runs as a privileged user, not
the user, and so the user needs to be able to modify the user's files.
The directory already should have the correct permissions, but no the
files. Fix those.

This let's us start guestfs guests on debian.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>