gitr: Ensure each fstype gets its own default settings
When changing the file system type from "nfs" to "tmpfs", I noticed
that the mount options did not also change from "vers=4.2,sec=sys"
to "size=75%". This is because all of the file system Kconfig
settings use the same Kconfig variable names.

Rename these so that the default settings for each file system
come up when changing the fstype to be tested.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>