workflows: Add an "ltp" workflow
LTP contains a number of file system-related tests, including the
nfslock tests, the fs* tests, and the syscall/*notify tests. These
tests have been a favorite of distributors' Q/A teams for many
years, so it makes sense to include them to help file systems
developers get in front of testing results that might be seen only
after an upstream kernel release is prepared for a Linux
distribution.

LTP test subgroups are enabled by Kconfig options to avoid running
tests that do not directly pertain to the developer's interests.

In "dedicated" mode, a separate target node is brought up for each
each test group, allowing multiple test groups to be run
concurrently. This helps increase the number of tests that can be
run in a given amount of elapsed time.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>