guestfs_q35: use hex for pci addr
Since we are prepending as a hex value ensure the input is converted
as such. This creates no functional changes but if we later want to
increase the drives beyond 9 then this will ensure we use the address
space appropriately. This will become more important later when we
add support for large-io experimentation to guestfs which does add
tons of drives onto a guest, enough to be able to experiment with
LBS support on different types of drives all in one system.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>