guestfs: use macros for drives for aarch64
Now that we have macros for drive generation for the XML files
we can share this with aarch64. So move all the code to generate
drives into a shared file and include it for both q35 and aarch64.

This then also augments aarch64 to also support large IO experimentation
(CONFIG_QEMU_ENABLE_EXTRA_DRIVE_LARGEIO)

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>