bringup: match default distro to user's distro
We've had kconfigs/Kconfig.distro for a while which let's us pick
up on the distro the user of kdevops is using but haven't used it
extensively. Heavy users of kdevops want sensible defaults so you
have to do less configuration of kdevops, one of the things we could
do to make things smoother is match the target distro to use with the
user's distro. So do that and be consistent over vagrant / guestfs and
promote the same best practice when using terraform to use the cloud.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>