guestfs: remove explicit tap0 device name
If you create more than one guest each guest will have the same tap0
device name, and when you try to start the second guest it will fail
due to the clash. We can avoid this by just not specifying the target
host name and letting libvirt figure it out. This does mean a slight
functional change in that the host will get something like vnet+1 id
instead of tap0. This fixes getting networking when using multiple
guests as when testing with fstests against a filesystem.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>