destroy_guestfs.sh: remove known ssh key
When destroying a guest we should remove the known key so that if
we re-create the guest we won't be asked about a new key.

Without this the ssh-injection will fail on a guest if an old key was
already used in the bringup proces with:

virt-sysprep -a $ROOTIMG --hostname $name --ssh-inject <...>

Fix this.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>