guestfs: verify new line on ssh include directive
If the ansible task added the include directive for kdevops and later a new
host entry was added (say with Vagrant), it means the Include directive is
followed by an entry without a new line. This will mean ssh won't use
that include file.

So we need to be a bit paranoid with this effort. So we are going to
first check if this sanity check was done first by looking for a special
new tag we're going to add now, if that exists we know we've our job and
can bail. Otherwise we're going to remove the old stale line, move it to
the top and ensure its at the top of the file. To ensure a new line is
used we use the ansible block module, and we take advantage of this by
adding the version of kdevops we use to add this. That's our marker that
the include directive is OK.

Fixes: e9390b898f98 ("guestfs: add the Include directive to ~/.ssh/config")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>