guestfs: add pcie passthrough XML generation / attach support
Most of the hard work to support PCIE passthrough support was already
done before for vagrant and through 'make dynconfig' for the dynamic
Kconfig stuff. This let's associate a PCIE device to a guest or lets us
pick which device onto which guest to peg a device to.

The only thing left to do for guestfs was to generate the XML files and
then attach them. When and how we do this was quite tricky though.

We just generate an XML file with the input source information, we do
this at 'make' time.

CXL folks may want to pay attention to this, it means you can use
XML output files for each target guest, or you dynamically generate
guests and their respective PCI topology at bring up time later.

To attach we just wait after we define the guest. As soon as we
define it, we can virsh attach devices to it.

To simplify things further, instead of assuming we know where exactly
on the guest we want to put a device we let libvirt figure that out.

This proves to be super clean and removes all the PCI passthrough
stuff from the XML input source file. This should also mean the CXL
dynamic topology we may want to later add support for could all be
kept out of the file source file and instead consist of XML input
files post device definition. All we'd really need for the XML device
to support CXL is to define the PCI-E reqs before we virsh define each
guest, and that is pretty generic.

Screenshot:

kdevops@gdemo ~ $ sudo nvme list
Node                  Generic               SN                   Model                                    Namespace Usage                      Format           FW Rev
--------------------- --------------------- -------------------- ---------------------------------------- --------- -------------------------- ---------------- --------
/dev/nvme0n1          /dev/ng0n1            S332NXAH300376       SAMSUNG MZQLW1T9HMJP-000ZU               1         119.93  GB /   1.92  TB      4 KiB +  0 B   CXV8301Q

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>