guestfs: wait for connections early
If we kick off the train of ansible tasks right away
we end up timing out. The right thing to do is to wait for
connections even before we try to gather facts. So be super
explicit about that for guestfs. This will save us tons of
silly odd timeout bug reports.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>