nfsd: add a pipefs-directory config to nfs.conf
Debian uses /run/rpc_pipefs for the pipefs mountpoint by default, but
the nfs.conf that we deploy doesn't have a pipefs-directory
configuration.

That will cause any userspace daemons that use the pipefs to fail to
start/restart because they'll now be looking for the pipefs to be
mounted on /var/lib/nfs/rpc_pipefs... unless we reboot the nfsd server
after we overwrite Debian's default nfs.conf with our own.

So let's just add the pipefs-directory configuration to nfs.conf
instead.

Signed-off-by: Scott Mayhew <smayhew@redhat.com>
Reviewed-by: Jeff Layton <jlayton@kernel.org>
Reviewed-by: Luis Chamberlain <mcgrof@kernel.org>