bringup: make guestfs default
guestfs has been tested in quite a bit of workflows already and other
than CXL and ZNS support, it's all there. ZNS support should be easy
to add, and we need to revisit how we scale CXL support for more
random dynamic topologies anyway, so that work can be done with
guestfs. So just make guestfs default now.

Acked-by: Jeff Layton <jlayton@kernel.org>
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>