gen_hosts/gen_nodes: clean up nfsd-related stuff
First, instead of checking if nfsd_threads is defined, create new
variable kdevops_nfsd_enable and use that.

Second, if nfsd is not enabled, then there shouldn't be "nfsd" and
"nfsd:vars" stanzas in the hosts file, so adjust the placement of the
conditionals in the templates accordingly.

No change in behavior.

Signed-off-by: Scott Mayhew <smayhew@redhat.com>
Reviewed-by: Jeff Layton <jlayton@kernel.org>
Reviewed-by: Luis Chamberlain <mcgrof@kernel.org>