gitr: Clone a fixed version of git
The current default checkout-target for git is "maint", but tests
with that release recently started failing. Adopt the same tactic
used for other workflows: pick a fixed known-working version of
git to clone for the gitr workflow.

This version will, of course, need to be updated from time to time.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>