devconfig: Copy timezone setting to target nodes
We want the time and timezone settings on target nodes to match the
hypervisor / control host settings so that things like journald,
watchdogs, and NFS/Kerberos work correctly.

Instead of changing the hypervisor's timezone setting, which is
somewhat surprising behavior, add a Kconfig option to simply set
each target node's timezone to be the same as the control host's.

Note: the systemd-timesyncd infrastructure assumes a modern version
of the timedatectl command is available on the control host and
target nodes. It might not be available on older Linux releases.

Since DEVCONFIG_ENABLE_SYSTEMD_TIMESYNCD_TIMEZONE results in
surprising behavior, its default setting is changed to N.

Reviewed-by: Luis Chamberlain <mcgrof@kernel.org>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>