workflows: add pre-fill experimentation script to help optimize preconditioning
We want to eventually add support for pre-conditioning drives so we
can run performance analysis on them. Part of pre-conditioning involves
doing a drive pre-fill, to write to every LBA. The typical approach is to
do the work sequentially and with just one thread with a larger iodepth
set. For high end capacity drives where capacity > 4 TiB on flash storage,
it begs the question if we can do better.

We want to verify if a single fio job on a modern system with an iodepth
matching an NVMe drive's (/sys/block/nvme0n1/device/queue_count) *might*
pre-fill a driver just as fast as 128 jobs with iodepth>=1 when using
the drive's optimal bs (/sys/block/nvme0n1/queue/optimal_io_size).
That is, we want to verify if single CPU thread on a modern system
should be able to saturate a full NVMe drive today then. We provide an
example script which can be used to verify / test this.

What we use by default would take the below attributes into
consideration:

blocksize  jobs   iodepth   time
128k       1      1         a
128k       1      129       b
128k       127    1         c
128k       127    129       d

Details:

We are used to doing pre-fill with something sequential and without
parallelization like this:

fio --filename=/dev/nvme0n1 -direct=1 -name drive-pre-fill \
	--readwrite=write --ioengine=io_uring \
	--blocksize=131072 \
	--iodepth=129

Unfortunately it is not easy to accomplish with fio pre-fill with
parallelization when using direct IO due to the alignment requirements,
and the offsets need to align for each job. Take an example drive
of about 4 TiB with 3840755982336 bytes exactly using a 4k LBA format as
an example. If you have 128 cores dividing the space up evenly gets you
30005906112 bytes for each thread but that value is not evenly
divisible by 4k or even 8k:

30005906112/4096
7325660.67187500000000000000

30005906112/(4096*2)
3662830.33593750000000000000

The right approach then is in the end dividing the space by evenly
divisible aligned blocks, and then figuring out how many blocks
of those your target block size uses, and then dividing that space
by the number of jobs you want to work with. You may or may not
end up with a remainder, in which case you will need a thread which
deals with the last remaining bytes.

Addressing support for block devices and target truncated files also
varies, and since we're growing support for large atomics in the kernel
it makes sense to start to untangle this mess and help with some sort
level of documentation for how to best take advantage of some of these
things

This adds an example prefill script for block device and files
which will gives us the fio commands required to run with optimal
settings given your drive's arbitrary capacity and arbitrary number of
jobs you specified.

By default, if a block device is used the drive's optimal-io in sysfs
for example /sys/block/nvme0n1/queue/optimal_io_size is used. This
is different to the value obtained in 'stat --format="%o"' and, and
so sadly the documentation for the value is not accurate. The value
returned is not the optimal IO size for block devices or files. We
should fix that documentation.

An enhancement done with our pre-fill is also leveraging the physical
block size instead of the logical block size when possible so to avoid
read-modify-writes. Note that there is no current way to obtain a
value such as the NPWG for drives if they don't have parity with a
large atomic (NAWUPF). So we have no current way for userspace to easily
know that a block device should be written to with a larger value than
the logical block size if it does not have a large atomic. The only
current way is for userspace to be NVMe-smart and know to align to the
NPWG on its own.

Example use cases:

prefill-fio-jobs.sh --target /dev/nvme0n1 --blocksize 32768 --verbose
target: /dev/nvme0n1
pbs: 4096
size: 3840755982336
bs: 32768
jobs: 128

aligned-pbs-blocks: 937684566
aligned-bs-per-block: 8

aligned-blocks: 117210570
aligned-block-jobs: 922917
unaligned-block-jobs: 111

aligned-jobs: 127
unaligned-jobs: 1

aligned-bytes-per-job: 30242144256
total-aligned-bytes: 3840752320512
remainder--block-bytes: 3661824
total-aligned-bytes + remainder-block-bytes == size ------> OK!

Fio command for all aligned data at blocksize 32768 using 127 jobs:

fio --filename=/dev/nvme0n1 -direct=1 -name drive-pre-fill-aligned-to-bs --readwrite=write --ioengine=io_uring --group_reporting=1 --alloc-size=1048576 --numjobs=127 --offset_increment=30242144256 --size=30242144256 --blocksize=32768

Fio command for all remaining data which needs to be issued at 4096:

fio --filename=/dev/nvme0n1 -direct=1 -name drive-pre-fill-aligned-to-pbs --readwrite=write --ioengine=io_uring --group_reporting=1 --offset=3840752320512 --size=3661824 --blocksize=4096

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>