provision: generalize bridge-ip further
We only provide the bridge IP on Kconfig if we're using vagrant or
guestfs, and we wrap some kconfig logic around this. We want to expand
on this and use the default bridge IP for mirroring as well, so that we
don't phone home to the internet when we have a git tree for local
trees. Today we only use git trees for the large trees for the Linux
kernel and qemu but we want to expand on that, and add all of the
mirrors a node fetches to a locall mirror.

To enable more wide spread use of the bridge IP we need to just make
it available on Makefiles, so we can take advantage of the fact that
although kconfig does not let us append variables for example we can't
use "default $(foo)/$(bar)" we *can* as the shell to append scripts for
us, we just need to promote the variables we need to shell.

We provide a simple scripts/append-makefile-vars.sh to make emphasis
that you need these variables on Makefiles so you can use them on
Kconfig.

With the bridge IP for all worlds defined we can now grow its use later
for allowing us to extend how we use bridging using the git protocol on
the local network. We'll address this next.

This makes no functional changes, it just let's us later use the
bridge IP variables on shell Kconfig magic.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>