mirrors: enable kdevops mirror
We recently added support for mirroring kdevops but didn't enable it
yet as we needed some more work, the work to get the bridge IP in a
clean way. Now that previous commits did the work for that enable
the mirror to be used, but only if we detect that the local server
has a copy of that mirror. The reason we check for the mirror is
that prior mirroring attempts may have been made with older versions
of kdevops and they would not have set up the kdevops mirror yet.
We therefore only enable the kdevops mirror to be used if and only if
the localhost has it present.

This is going to save us tons of bandiwdth, as each guest when testing
filesystems was git cloning kdevops and kdevops already is about over
200 MiB.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>