mirrors: add kdevops git alternative for fstests
kdevops supports its own CONFIG_GIT_ALTERNATIVES which allows
kdevops to mirror upstream projects on its own organization, so
that our own fork can host patches which perhaps have not landed
upstream on fstests yet, and also to ensure reproducible workflows.
If we follow upstream, the branch may be a moving target and sometimes
thigns may work sometimes things will not. For deatils refer to our
Kconfig documentation of GIT_ALTERNATIVES.

This now mirrors the fstests kdevops git alternatives. We will place
these mirrored alternatives under a directory representing the
alternative source, in this case its kdevops, /mirror/kdevops/fstest.git.
Note that we have an alternative for fstests for btrfs but this doesn't
yet properly pronounce itself as an alternative itself, its was just
caked in as another option. We should fix that so that we can also
mirror btrfs's fstests and be able to easily use it.

Now with today's defaults you will end up using mirroring for
both kdevops and fstests when you enable the fstests workflow
if a mirror was found for each one on your system:

grep git extra_vars.yaml | grep git
kdevops_git: git://192.168.122.1/mirror/kdevops.git
kdevops_git_version: HEAD
fstests_git: git://192.168.122.1/mirror/kdevops/fstests.git
fstests_git_version: HEAD

It is important to use --reference here as the kdevops fork and so
can share objects on the server with the fstests tree. If you add
support for a tree which is a fork, do the same.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>