mirrors: use blktests mirror when its available
There isn't any blktests mirror out there yet, and we also don't even
have a form of it on kdevops since it is a rather slow moving target
compared to fstests and does not break as often. So mirroring for it
is pretty straight forward, and the only difference is the IP address
betwen the bridge for vagrant and guestfs. Enable the mirror if the
localhost is detected to use mirroring.

After enabling the blktests workflow and running make we end up
with:

grep blktests_git extra_vars.yaml
blktests_git: git://192.168.122.1/mirror/blktests.git

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>