mirrors: add new mirrors.yaml and add xfsprogs
The process to add mirrors is pretty error prone. A lot of it
is because we're doing it the old way of shell scripts to scale
and we can do better. Just use kconfig to define preferences,
ansible to run tasks, and define a yaml file with the mirror
preferences outlined based on the target variability the user
has opted into.

We no longer have to touch the linux-mirror-systemd/Makefile
with clunky stuff, all we need to do to add new mirrors is edit
playbooks/roles/linux-mirror/templates/mirrors.yaml.j2 and ensure
we have sensible defaults. This also lets us do variability.

The entire process to install user units and timers, and also
clone the mirrors is now done with python scripts that read the
input yaml file.

Also make use of this now in one shot. It will be clearer how
to add new mirrors in one shot in the next commit where we don't
add the infrastructure for it.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>